86cym3xfh: Re-structure code into client & server folders to prevent cross-imports
