+++ app/client/ui-modules/hooks/useImagePreload.ts
