+++ app/client/ui-modules/property-listing/components/ListingPropertyDetails.tsx
