+++ app/client/ui-modules/theming/icons/CalendarIcon.tsx
