+++ app/client/ui-modules/theming/icons/LeftCircularArrowIcon.tsx
