+++ app/client/ui-modules/theming/icons/RightCircularArrowIcon.tsx
