+++ app/server/model/TaskDTO.ts
@@ -1,0 +1,5 @@
+
+export type TaskDTO = {
+  taskId: string,
+  text: string
+}
